pyproject.toml
[project]
name = "lever-mcp"
version = "0.1.0"
description = "MCP server for Lever.co"
dependencies = [
    "fastmcp",
    "httpx",
    "pydantic",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lever_mcp"]
